+++ src/state.c
@@ -81,7 +81,7 @@ void ss_grid_init(scene_state_t *ss) {
         ss->grid.group[i].fader_max = 16383;
     }
     
-    for (u8 i = 0; i < GRID_BUTTON_COUNT; i++) {
+    for (u16 i = 0; i < GRID_BUTTON_COUNT; i++) {
         ss_grid_common_init(&(ss->grid.button[i].common));
         ss->grid.button[i].latch = 0;
         ss->grid.button[i].state = 0;
